requirements.txt
boto3==1.10.19
Flask==1.1.1
neuralgym==0.0.1
numpy==1.17.4
Pillow==6.2.1
PyYAML==5.1.2
tensorflow==1.15.0
torch==1.3.1
torchvision==0.4.2
